requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
ngrok==0.6.0
Pillow==9.5.0
# psycopg2==2.8.1
pyngrok==6.0.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
